# contracts/openapi.v1.yaml
openapi: 3.1.0
info:
  title: AI Transaction Authorization Infrastructure API
  version: "1.0.0"
  summary: Deterministic APPROVE/DENY/REVIEW for AI-initiated transactions, before they reach the payment provider.
  description: |
    The decision-plane HTTP surface. A caller POSTs a transaction to `/v1/authorize`
    **before** invoking its payment provider and receives a deterministic, explainable,
    Ed25519-signed decision. **This service never moves, holds, or settles money.**

    ## Two distinct signatures — do not conflate them (ROADMAP A#1)
    - **Request authentication (caller → us): HMAC, symmetric.** Every request carries
      `X-Api-Key`, `X-Signature` (HMAC over a canonical payload), `X-Nonce`, and
      `X-Timestamp`. The caller holds the key; symmetric MAC is appropriate. See the
      `ApiKeyAuth` + `HmacSignature` security schemes below.
    - **Decision signature (us → everyone): Ed25519, asymmetric.** Each `Decision`
      carries a `signature` object produced with our Ed25519 private key. Anyone —
      customer, auditor, regulator — verifies it against the PUBLISHED public key
      (`GET /v1/keys/public`) without holding a secret. HMAC here would let a key-holder
      forge our decisions and is forbidden.

    ## Determinism (ROADMAP A#7)
    A decision is a pure function of `(request, policy_version, evaluated_at, counter_snapshot)`.
    `evaluated_at` and `counter_snapshot` are injected by the orchestrator; predicates never
...
  contact:
...
  license:
...
servers:
...
security:
  - ApiKeyAuth: []
    HmacSignature: []
paths:
  /v1/authorize:
    post:
      tags: [authorize]
      operationId: authorize
      summary: Evaluate a transaction and return a signed decision.
      description: |
        Evaluates the request against the caller's active policy bundle and returns an
        APPROVE / DENY / REVIEW decision with a structured explanation and an Ed25519
        signature. Idempotent per `idempotency_key`: replaying an identical request with
        the same key returns the same decision and creates no additional side effect.
      parameters:
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "./schemas/authorize-request.schema.json"
            examples:
              vendorPayment:
                summary: Procurement agent paying a vendor
                externalValue: ./examples/authorize-request.example.json
      responses:
        "200":
          description: A decision was produced (APPROVE, DENY, or REVIEW — all are 200).
          headers:
            X-Decision-Id:
              description: The decision_id, mirrored for logging/correlation.
              schema: { type: string }
            X-Shadow:
              description: |
                Present and "true" when the decision is advisory (a shadow-mode key):
                the caller must NOT enforce it. Mirrors the signed `shadow` field.
              schema: { type: string, enum: ["true"] }
          content:
            application/json:
              schema:
                $ref: "./schemas/decision.schema.json"
              examples:
                approve:
                  summary: APPROVE
                  externalValue: ./examples/decision-approve.example.json
                deny:
                  summary: DENY (monthly budget exceeded)
                  externalValue: ./examples/decision-deny.example.json
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/ReplayDetected"
        "422":
          $ref: "#/components/responses/Unprocessable"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/authorize/{id}/capture:
    post:
      tags: [authorize]
      operationId: captureHold
      summary: Commit the budget hold placed by an APPROVE (two-phase, A#2).
      description: |
        Call after the downstream payment succeeds. Commits the reservation the
        matching authorize placed (the decision carried a `capture_within` obligation
        whose `hold_ref` is this id). Idempotent. Rejected 409 if the hold was voided or
        expired; 404 if there is no hold for the id.
      parameters:
        - $ref: "#/components/parameters/DecisionId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      responses:
        "200":
          description: The hold state after capture.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HoldState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/authorize/{id}/void:
    post:
      tags: [authorize]
      operationId: voidHold
      summary: Release the budget hold placed by an APPROVE (two-phase, A#2).
      description: |
        Call if the downstream payment fails or is aborted. Releases the reservation.
        Idempotent. Rejected 409 if the hold was already captured; 404 if there is no
        hold for the id. An untouched hold also auto-releases at its TTL.
      parameters:
        - $ref: "#/components/parameters/DecisionId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      responses:
        "200":
          description: The hold state after void.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HoldState" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /v1/decisions/{id}:
    get:
      tags: [decisions]
      operationId: getDecision
      summary: Fetch a stored decision record with verification status.
      parameters:
        - name: id
          in: path
          required: true
          description: The decision_id (ULID).
          schema:
            $ref: "./schemas/common.schema.json#/$defs/ULID"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      responses:
        "200":
          description: The stored decision, with `record_hash` and read-time `signature_verified`.
          content:
            application/json:
              schema:
                $ref: "./schemas/decision.schema.json"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/decisions:
    get:
      tags: [decisions]
      operationId: listDecisions
      summary: List stored decisions (paginated, filterable).
      parameters:
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
        - name: agent_id
          in: query
          required: false
          schema: { type: string }
        - name: verdict
          in: query
          required: false
          schema: { type: string, enum: [APPROVE, DENY, REVIEW] }
        - name: from
          in: query
          required: false
          description: Inclusive lower bound on evaluated_at (RFC 3339).
          schema: { type: string, format: date-time }
        - name: to
          in: query
          required: false
          description: Exclusive upper bound on evaluated_at (RFC 3339).
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          required: false
          description: Opaque pagination cursor from a previous response.
          schema: { type: string }
      responses:
        "200":
          description: A page of decisions.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "./schemas/decision.schema.json"
                  next_cursor:
                    type: [string, "null"]
                    description: Cursor for the next page, or null when exhausted.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/policies:
    post:
      tags: [control-plane]
      operationId: createPolicy
      summary: Create a draft policy.
      description: |
        Creates a draft policy owned by the caller's org (taken from the API key, never
        the body). Rules are validated authoritatively at publish, so a draft may be a
        work in progress.
      parameters:
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyWriteRequest" }
      responses:
        "201":
          description: The created draft policy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/policies/active:
    get:
      tags: [control-plane]
      operationId: getActivePolicy
      summary: The org's currently serving policy version.
      description: |
        Returns the version the org's decision plane is serving (ROADMAP A#4). Reads
        control-plane state, so it reflects a publish immediately; the decision plane
        converges to it within the bundle-refresh window.
      parameters:
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      responses:
        "200":
          description: The active bundle pointer.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BundleRef" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/policies/{id}:
    get:
      tags: [control-plane]
      operationId: getPolicy
      summary: Fetch a policy's working copy.
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      responses:
        "200":
          description: The policy's current working copy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [control-plane]
      operationId: updatePolicy
      summary: Replace a policy's working-copy fields.
      description: |
        Edits the draft (name, agents, rules). Does not publish and does not change the
        active version — the previously published version keeps serving until the next
        publish.
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyWriteRequest" }
      responses:
        "200":
          description: The updated working copy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/policies/{id}/publish:
    post:
      tags: [control-plane]
      operationId: publishPolicy
      summary: Validate, version, sign, and publish a policy.
      description: |
        Validates the working copy, mints an immutable content-hash version
        (`polv_<sha256>`), Ed25519-signs it, stores it, and designates it the org's
        active bundle. Republishing byte-identical content is idempotent (same hash).
        An invalid policy is rejected 422 with field-level problems.
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PublishRequest" }
      responses:
        "200":
          description: The published, signed version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyVersion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /v1/policies/{id}/rollback:
    post:
      tags: [control-plane]
      operationId: rollbackPolicy
      summary: Restore a prior version by re-pointing the active bundle.
      description: |
        Re-points the org's active bundle at an existing version hash and resets the
        working copy to that version's content (TRD §6). Content-addressed and
        immutable, so no new version is created.
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RollbackRequest" }
      responses:
        "200":
          description: The reinstated version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PolicyVersion" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/policies/{id}/versions:
    get:
      tags: [control-plane]
      operationId: listPolicyVersions
      summary: List a policy's immutable versions (newest first).
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      responses:
        "200":
          description: The policy's versions, newest first.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/PolicyVersion" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/policies/{id}/simulate:
    post:
      tags: [control-plane]
      operationId: simulatePolicy
      summary: Dry-run a policy version (or the draft) against a batch of requests.
      description: |
        Evaluates each supplied request against a target policy version (`version_hash`)
        or, when omitted, the current working-copy draft. Returns per-request would-be
        decisions — all marked `shadow: true`, UNSIGNED, and never persisted. A request's
        own `evaluated_at` is honored (replaying historical traffic).
      parameters:
        - $ref: "#/components/parameters/PolicyId"
        - $ref: "#/components/parameters/ApiKey"
        - $ref: "#/components/parameters/Signature"
        - $ref: "#/components/parameters/Nonce"
        - $ref: "#/components/parameters/Timestamp"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SimulateRequest" }
      responses:
        "200":
          description: Per-request would-be verdicts.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimulateResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/Unprocessable" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        Per-key API credential, prefix `azn_live_` / `azn_test_`. Only `sha256(key)` is
        stored server-side; comparison is constant-time (TRD §11).
    HmacSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: |
        HMAC (symmetric) over the canonical request payload using the API key secret, paired
        with `X-Nonce` and `X-Timestamp`. Verified constant-time; timestamp skew beyond the
        configured window and reused nonces are rejected. This authenticates the CALLER and is
        entirely separate from the Ed25519 signature on the returned Decision.
  parameters:
    ApiKey:
      name: X-Api-Key
      in: header
      required: true
      schema: { type: string, pattern: "^azn_(live|test)_[A-Za-z0-9_-]+$" }
    Signature:
      name: X-Signature
      in: header
      required: true
      description: HMAC-SHA256 of the canonical payload, hex or base64url.
      schema: { type: string }
    Nonce:
      name: X-Nonce
      in: header
      required: true
      description: Unique per request; cached per key to reject replays.
      schema: { type: string, minLength: 8, maxLength: 128 }
    Timestamp:
      name: X-Timestamp
      in: header
      required: true
      description: RFC 3339 request time; rejected if skew exceeds the configured window.
      schema: { type: string, format: date-time }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Optional transport mirror of the body `idempotency_key`. When both are present they
        must match. Makes authorize retry-safe.
      schema: { type: string, minLength: 8, maxLength: 128 }
    PolicyId:
      name: id
      in: path
      required: true
      description: The policy id (`pol_<ULID>`).
      schema: { type: string, minLength: 1 }
    DecisionId:
      name: id
      in: path
      required: true
      description: The decision id (ULID) whose budget hold is being captured/voided.
      schema: { $ref: "./schemas/common.schema.json#/$defs/ULID" }
  responses:
    ValidationError:
      description: Malformed request (schema/shape).
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Unauthorized:
      description: Missing/invalid API key or bad HMAC signature.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Forbidden:
      description: Key revoked or lacks scope.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    NotFound:
      description: No such resource.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    ReplayDetected:
      description: Nonce already seen for this key.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Unprocessable:
      description: Well-formed but semantically invalid (e.g. currency/amount inconsistency).
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    RateLimited:
      description: Per-key rate or burst limit exceeded.
      headers:
        Retry-After:
          description: Seconds to wait before retrying.
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Unavailable:
      description: Decision plane could not serve (e.g. no last-known-good bundle). Never a silent guess.
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
    Conflict:
      description: Illegal state transition (e.g. capturing a voided or expired budget hold).
      content:
        application/problem+json:
          schema: { $ref: "./schemas/error.schema.json" }
  schemas:
    HoldState:
      type: object
      description: A budget hold's state after a capture/void transition (Task 3.1).
      additionalProperties: false
      required: [decision_id, state]
      properties:
        decision_id: { type: string, description: The decision id the hold is addressed by. }
        state:
          type: string
          enum: [held, captured, voided, expired]
          description: Lifecycle state. captured/voided/expired are terminal.
        budget_id: { type: string }
        amount: { $ref: "./schemas/common.schema.json#/$defs/MonetaryAmount" }
        currency: { $ref: "./schemas/common.schema.json#/$defs/CurrencyCode" }
        expires_at: { $ref: "./schemas/common.schema.json#/$defs/Timestamp" }
    PolicyWriteRequest:
      type: object
      description: |
        Body for create/update. The org is taken from the API key, never the body; rules
        reuse the frozen policy rule shape. Validated authoritatively at publish.
      additionalProperties: false
      required: [name, rules]
      properties:
        name: { type: string, minLength: 1, maxLength: 256 }
        agents:
          type: array
          description: Agents this policy targets; empty/omitted = every agent in the org.
          items: { type: string, minLength: 1, maxLength: 128 }
        rules:
          type: array
          minItems: 1
          items: { $ref: "./schemas/policy.schema.json#/$defs/Rule" }
    Policy:
      type: object
      description: A policy's mutable working copy plus its active-version pointer.
      additionalProperties: false
      required: [id, org_id, name, agents, rules, status, created_at, updated_at]
      properties:
        id: { type: string, description: "pol_<ULID>." }
        org_id: { type: string }
        name: { type: string }
        agents:
          type: array
          items: { type: string }
        rules:
          type: array
          items: { $ref: "./schemas/policy.schema.json#/$defs/Rule" }
        status: { type: string, enum: [draft, published] }
        active_version_hash:
          type: string
          description: The version currently serving; absent until the first publish.
        created_at: { $ref: "./schemas/common.schema.json#/$defs/Timestamp" }
        updated_at: { $ref: "./schemas/common.schema.json#/$defs/Timestamp" }
...
    BundleRef:
      type: object
      description: The per-org active-bundle pointer (which published version is serving).
      additionalProperties: false
      required: [org_id, policy_id, version_hash, updated_at]
      properties:
        org_id: { type: string }
        policy_id: { type: string }
        version_hash: { type: string }
        updated_at: { $ref: "./schemas/common.schema.json#/$defs/Timestamp" }
    PublishRequest:
      type: object
      additionalProperties: false
      properties:
        author:
          type: string
          description: Optional provenance label; defaults to the authenticated key id.
    RollbackRequest:
      type: object
      additionalProperties: false
      required: [version_hash]
      properties:
        version_hash: { type: string, description: The prior version hash to reinstate. }
...